refactor: implement cogs
