+++ utils/config.py
@@ -1,0 +1,5 @@
+import json
+
+def load_config(path = "config/config.json"):
+    with open(path, "r", encoding = "utf-8") as f:
+        return json.load(f)
